Initial Functionality (board capability + DIO)
This commit provides initial functionality for the a-star robot
controller board with Firmata, including implementation of the system
capability query (needed for handshake), and digital output.

This particular version of the firmware also indiscriminately sends
analog values back to the host. This is a bug that will be fixed shortly
(see comment in loop())