Implement motor watchdog
This commit implements a motor watchdog in the form of heartbeat
messages. The controller watches for "hbon"/"hboff" messages (sent via Firmata
STRING_DATA messages) and uses that to enable/disable the watchdog.

In the main loop, the controller watches for "hb" messages from the host
and uses that to feed the watchdog. If the watchdog is not fed in time,
the motors shut off.